Several bugs fixed in search_grid. Search_grid is now used in the main program and can be turned on and off by a new config option SEARCH_GRID. In addition, all parameters for constructing the search grid can be set in the config file (names start with SG_...).
Currently the power_law_exponent is set to 2 for efficiency reasons (pow
replaced by multiplication) and cannot be changed by the config option
SG_POWER_LAW_EXPONENT. However, so far the speed gain is minimal, and the
old behaviour might be re-implemented after some testing in the near
future.

The total number of bisections is calculated and printed for testing and
comparing the performance for different search grid parameters.



git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/trunk@372 46abec54-52ec-0310-8d7b-a975267b85f2