BugFixes: The potential of the central black hole (cenma) was essentially added twice in calculate_potential. It is already contained in the recursive formula by adding cenma.m to Mtotal. Commented out the corresponding loop.
Changed star.phi[0] such that it is an extension of the recursion formula, 
identifying particle 0 as the central black hole with star[0].r=0 and 
m= cenma.m . The terms cenma.m/r[0] have been neglected (BH self-gravity). 
If cenma.m== 0 star[0].phi==star[1].phi as before. 

Corrected the function "potential" for the case star[0].r==0 < r < star[1].r .
Previously, star[1].phi was returned, which is only correct for cenma.m==0, as
star[0].phi== star[1].phi. Added it as an extra case, since the interpolation 
formula cannot be directly applied because of the 1/star[0].r terms.

Added the contribution of the central black hole (particle 0) to the total
potential energy in ComputeEnergy and RecomputeEnergy.



git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/trunk@385 46abec54-52ec-0310-8d7b-a975267b85f2