Added extensive checks that test if the results as well as intermediate results become infinite or NaNs.
Changed the radicand (essntially 1./vr) such that it has the form as FUNC if
debugging is enabled.

The if-else-if statements that check whether the point is within the last, or 
first interval now include the interval endpoints.

Removed the calculation of the interval where a star with radius r is, as this
is now handled by calc_orbit_rs. kmax and kmin are now part of the orbit_rs 
struct.


git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/trunk@420 46abec54-52ec-0310-8d7b-a975267b85f2