Added initial support for CUDA library.
Also cleaned up the configure.ac a little. Care must be taken when linking,
as the cudart library is only available as shared library. So static
linking is disabled when --enable-cuda is set, until I figured out a way
to only enable shared linking for libcudart.

Another problem is how to pass gcc's -malign-double to other compilers
like icc on x86 platforms.



git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/branches/cmc-autotools@763 46abec54-52ec-0310-8d7b-a975267b85f2