Starting new branch for addition of three-body binary formation.
New branch, cmc-3bbformation, for changes related to three-body binary formation. Many bugs to be fixed, but it does compile.

My additions were made on revision 708, but I have merged the more recent additions to the repository cmc-springcleaning branch up to revision 775. Note, in bse_wrap/Makefile, I have removed "popsynth-mpi" in line 14, used when making all, because the compiler cannot find mpicc; so here the mpi version of popsynth is not used.


git-svn-id: svn+ssh://svn.wcas.northwestern.edu/var/svn/cmc/branches/cmc-3bbformation@779 46abec54-52ec-0310-8d7b-a975267b85f2