Analytic Peters inspiral + Smbh compute energy fixes (#43)
* some stuff

* Analytic Peters inspiral for e->1

Refined peters_t_insp to use analytic form of inspiral time for the e->1 case.  From a software perspective this meant defining a custom gsl error handler to interact with this use of gsl_integration_qags.  The 0.99 lower limit precision was arbitrarily chosen.

* Correction to SMBH Etotal.P calculation

The previous version had the correct expression for the SMBH buf_reduce[2] contribution (to keep up with the double counting in the non-SMBH calculation).  The issue was that because this was included in the loop over all objects it had to be divided by the number of objects (clus.N_MAX), but this operation was strikingly similar to the mass unit conversions in the previous lines (/ clus.N_STAR); it would appear that this similarity was what caused clus.N_STAR to be used instead, which meant the second buf_reduce[2]+= line was adding something too small by a multiplicative factor of clus.N_MAX/clus.N_STAR.  As the clusters lost objects, this bug eventually caused ComputeEnergy to incorrectly yield Etotal.tot>0, terminating the run.

tl;dr: Using the wrong global variable caused the diagnostic ComputeEnergy to think the cluster was unbound.  This was fixed, along with switching out the appropriate "/ clus.N_STAR"s for "* madhoc"s to hopefully avoid this error in the future.

* Missing semicolon

Co-authored-by: Carl Rodriguez <carllouisrodriguez@gmail.com>